requirements.txt
C:\Users\Yashvi\Hackathon>pip freeze 
absl-py==2.0.0
aiofiles==22.1.0
aiosqlite==0.19.0
altair==5.2.0
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
category-encoders==2.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.16.0
clusters==66.0.2
colorama==0.4.6
comm==0.1.3
confection==0.1.4
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.8
Cython==3.0.10
datefinder==0.7.3
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
dm-tree==0.1.8
easyocr==1.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
fsspec==2023.9.2
gast==0.5.4
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
imbalanced-learn==0.12.0
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jax==0.4.25
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.11
mistune==2.0.5
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
namex==0.0.7
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.2
ninja==1.11.1.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
ntscraper==0.3.12
numba==0.59.0
numpy==1.24.3
oauthlib==3.2.2
object-detection==0.0.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.5.0
plotly==5.20.0
pocketsphinx==5.0.2
pooch==1.8.1
preshed==3.0.9
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==15.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydot==2.0.0
pydotplus==2.0.2
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pyspark==3.5.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.12.25
requests==2.30.0
requests-oauthlib==1.3.1
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.19.0
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
srsly==2.4.8
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorflow==2.16.1
tensorflow-estimator==2.14.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
thinc==8.2.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.9.26
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.1
tqdm==4.66.1
traitlets==5.9.0
transformers==4.38.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
ultralytics==8.2.14
uri-template==1.2.0
urllib3==2.0.2
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.0
widgetsnbextension==4.0.7
wrapt==1.14.1
wsproto==1.2.0
y-py==0.5.9
ypy-websocket==0.8.2